fix: make ChunkOrder i64 data type to accept min sequence number 0 and match with data type of sequence number (#4888)
* fix: make ChunkOrder u64 data type to accept min sequence number 0

* fix: make ChunkOrder i64 to match with sequence number type

+++ querier/src/ingester/mod.rs
@@ -778,7 +778,7 @@ impl QueryChunk for IngesterPartition {
     fn order(&self) -> ChunkOrder {
         // since this is always the 'most recent' chunk for this
         // partition, put it at the end
-        ChunkOrder::new(u32::MAX).unwrap()
+        ChunkOrder::new(i64::MAX).unwrap()
     }
 }
